Make observability optional.
The observability stack takes on resources, when
running in a github runner. For that reason we can't
default to true, as it might impact in the test results.
We need to be careful when we enable it.
This PR makes the local execution via local-k8s.sh
running with observability by default, but it
disables it when running it via Github Actions.