Add support for the RBAC feature and --debug.
This PR adds two new options RBAC (bool) and AUTH_CONFIG
(string) which allow configuring RBAC, as well as authorization
and authentication in Weaviate.

It also improves the help information displayed when no
option is passed or when local-k8s.sh --help is executed.

Last, it improves the busy loops that wait for services to
be ready which were kind of broken before and adds an automatic
way to pass authentication if required.

And it adds a DEBUG option, useful for debugging issues when
running on a Github action, as it enables extra logs (set -x)

Includes also a new job that retrieves Weaviate's latest version
to run the local-k8s Github workflows on that latest version.