Verify that required ports aren't in use before deploying.
One of the biggest issues is that the local-k8s deployment
fails because it was forgotten that another weaviate instance
was already running.
This PR checks if the ports required by local-k8s are in use
already.
Increase defaultTimeout for observability stack, as it can't
default to 0 in latest versions.