Improve pod's port-forwarding to make it robust.
Make the kubectl forwarding process more robust and skip the check
for a port if it's being used by another process. This way we don't
block the CI, as currently 8084 is being used by the ubicloud runners.

Also display if a port isn't being forwarded due to being already in
use.